+++ CHANGELOG.md
@@ -2,3 +2,11 @@
 -
 # 0.2 cell merge update
 - For better readability, implemented cell merging on subjects, filters and actions if there are multiple entries.
+
+# 0.3 single contract feature update
+- indicate service graph on subject
+- simplification in src/dst port representation.
+- single contract support
+
+
+